add database, viewmodel, converter pour Utilisateur

+++ app/src/main/java/com/example/pllrun/Classes/UtilisateurDao.kt
@@ -21,6 +21,6 @@ interface UtilisateurDao {
     suspend fun delete(utilisateur: Utilisateur)
 
     @Query("SELECT * from utilisateur WHERE id = :id")
-    fun getItem(id: Int): Flow<Utilisateur>
+    fun getItem(id: Long): Flow<Utilisateur>
 
 }
